relay/pyproject.toml
[project]
name = "sieve-relay"
version = "0.1.0"
description = "Minimal capture relay for Neural Sieve"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "aiosqlite>=0.20",
    "argon2-cffi>=23.1",
    "pydantic-settings>=2.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
relay = "relay.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["relay"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
